Make Win32 better, part I

+++ include/platform/time_null.h
@@ -8,7 +8,7 @@
 static inline uint64_t get_time(void)
 {
 	clock_t res = clock();
-	uint64_t val = res * (1000000000L / CLOCKS_PER_SEC);
+	uint64_t val = (uint64_t)res * (1000000000L / CLOCKS_PER_SEC);
 	return val;
 }
